refactor: Enhanced user-saving function in local storage to prevent duplicates

+++ src/utils/localStorage.ts
@@ -2,20 +2,32 @@ import GitHubUser from "../Interfaces/GitHubUser";
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export async function addUserToLocalStorage(newUser: GitHubUser | undefined) {
-    try {
-      if(!newUser) return;
-      const storedData = await AsyncStorage.getItem('githubUserData');
-      const parsedStoredData = storedData ? JSON.parse(storedData) : [];
-      console.log(storedData)
+  try {
+    if (!newUser) return;
+
+    const storedData = await AsyncStorage.getItem('githubUserData');
+    const parsedStoredData = storedData ? JSON.parse(storedData) : [];
+
+    const existingUserIndex = parsedStoredData.findIndex((user: GitHubUser) => user.login === newUser.login);
+
+    if (existingUserIndex !== -1) {
+
+      parsedStoredData.splice(existingUserIndex, 1)
+
       parsedStoredData.push(newUser);
-  
       const updatedData = JSON.stringify(parsedStoredData);
       await AsyncStorage.setItem('githubUserData', updatedData);
-    } catch (error) {
-      console.error('Erro ao adicionar usuário no AsyncStorage:', error);
+    } else {
+      parsedStoredData.push(newUser);
+
+      const updatedData = JSON.stringify(parsedStoredData);
+      await AsyncStorage.setItem('githubUserData', updatedData);
     }
+  } catch (error) {
+    console.error('Erro ao adicionar/atualizar usuário no AsyncStorage:', error);
   }
-  
+}
+
   export async function getUsersFromLocalStorage(): Promise<GitHubUser[]> {
     try {
       const storedData = await AsyncStorage.getItem('githubUserData');
